WIP pre-commit
