+++ twiinit_demos/cpu/systems/__init__.py
@@ -1,11 +1,26 @@
 # Copyright (C) 2024, twiinIT
 # SPDX-License-Identifier: Apache2.0
 
+from .parametric_blade_geometry import ParametricBladeGeometry  # isort:skip
+
+from .casing_geometry import CasingGeometry
 from .cpu import CPU
 from .fan import Fan
 from .fancontroller import FanController
 from .heatexchanger import HeatExchanger
+from .rotor_geometry import RotorGeometry
 
+from .fan_geometry import FanGeometry  # isort:skip
 from .cpusystem import CPUSystem  # isort:skip
 
-__all__ = ["CPU", "CPUSystem", "Fan", "FanController", "HeatExchanger"]
+__all__ = [
+    "ParametricBladeGeometry",
+    "CPU",
+    "CPUSystem",
+    "Fan",
+    "FanController",
+    "HeatExchanger",
+    "CasingGeometry",
+    "FanGeometry",
+    "RotorGeometry",
+]
